+++ src/main/java/edu/harvard/iq/dataverse/api/Files.java
@@ -45,6 +45,7 @@ import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -531,7 +532,73 @@ public class Files extends AbstractApiBean {
                 .type(MediaType.TEXT_PLAIN) //Our plain text string is already json
                 .build();
     }
+    @POST
+    @AuthRequired
+    @Path("{id}/metadata/version/{datasetVersionId}")
+    public Response updateFileMetadata(@Context ContainerRequestContext crc, @FormDataParam("jsonData") String jsonData,
+                                       @PathParam("id") String fileIdOrPersistentId,
+                                       @PathParam("datasetVersionId") String datasetVersionId) {
+        DataverseRequest req;
+        try {
+            req = createDataverseRequest(getRequestUser(crc));
+        } catch (Exception e) {
+            return error(BAD_REQUEST, "Error attempting to request information. Maybe a bad API token?");
+        }
+        final DataFile df;
+        FileMetadata fm = null;
+        try {
+            df = execCommand(new GetDataFileCommand(req, findDataFileOrDie(fileIdOrPersistentId)));
+            fm = df.getFileMetadata();
+            for (FileMetadata md : df.getFileMetadatas()) {
+                if (datasetVersionId.equals(String.valueOf(md.getDatasetVersion().getId())) ||
+                        datasetVersionId.equals(md.getDatasetVersion().getFriendlyVersionNumber())) {
+                    fm = md;
+                }
+            }
+        } catch (Exception e) {
+            return error(BAD_REQUEST, "Error attempting get the requested data file.");
+        }
 
+        try {
+            OptionalFileParams optionalFileParams = null;
+            if (jsonData != null) {
+                // Load optional params via JSON
+                optionalFileParams = new OptionalFileParams(jsonData);
+            } else {
+                return error(BAD_REQUEST, "Missing Form Data!");
+            }
+            List<FileMetadata> fmdListMinusCurrentFile = new ArrayList<>();
+            for (FileMetadata fileMetadata : df.getFileMetadatas()) {
+                if (!fileMetadata.getDataFile().equals(df)) {
+                    fmdListMinusCurrentFile.add(fileMetadata);
+                }
+            }
+            jakarta.json.JsonObject jsonObject = JsonUtil.getJsonObject(jsonData);
+            String incomingLabel = jsonObject.getString("label", null);
+            String incomingDirectoryLabel = jsonObject.getString("directoryLabel", null);
+            String existingLabel = fm.getLabel();
+            String existingDirectoryLabel = fm.getDirectoryLabel();
+            String pathPlusFilename = IngestUtil.getPathAndFileNameToCheck(incomingLabel, incomingDirectoryLabel, existingLabel, existingDirectoryLabel);
+            if (IngestUtil.conflictsWithExistingFilenames(pathPlusFilename, fmdListMinusCurrentFile)) {
+                return error(BAD_REQUEST, BundleUtil.getStringFromBundle("files.api.metadata.update.duplicateFile", Arrays.asList(pathPlusFilename)));
+            }
+
+            optionalFileParams.addOptionalParams(fm);
+            execCommand(new UpdateDatasetVersionCommand(fm.getDataFile().getOwner(), req, fm));
+
+        } catch (DataFileTagException ex) {
+            return error(BAD_REQUEST, ex.getMessage());
+        } catch (Exception e) {
+            return error(Response.Status.INTERNAL_SERVER_ERROR, "Error updating metadata to DataFile: " + e);
+        }
+
+        String jsonString = fm.asGsonObject(true).toString();
+        return Response
+                .status(Response.Status.OK)
+                .entity("File Metadata update has been completed: " + jsonString)
+                .type(MediaType.TEXT_PLAIN) //Our plain text string is already json
+                .build();
+    }
     @GET
     @AuthRequired
     @Path("{id}")